fix(print): Fix lines height not correctly detected
Adding a margin on a element makes it complicated to get height +
margins because of collapsing borders. Even if we measure correctly the
cmLine with margins (like jQuery.outerHeight()), the problem is that the
margin is in the contained element (in that case the section title), and
thus not measured properly.

Solution is to use dislay: inline-block to
disable collapsing borders.